Refactored, seperated out the RSBC and FF related tables
